change speed data type, fix mistake that caused speed values to get truncated to 0 and result in a pwm of 1500

+++ src/navigation/include/navigation/rc_communicator.h
@@ -5,7 +5,7 @@
 #include "mavros_msgs/OverrideRCIn.h"
 
 void rc_raw(int channel, int pwm);
-void rc(int channel, int pwm);
+void rc(int channel, float pwm);
 enum RCCHANNEL {
   PITCH_CHNL,
   ROLL_CHNL,
